cast mask

+++ streamlit_app.py
@@ -140,10 +140,10 @@ else:
         if REQUIRE_SINGLE_EAGLE:
             # Only proceed if YOLO found exactly one eagle
             if n_objects == 1:
-                mask = masks[0].cpu().numpy() > 0.5
+                mask = (masks[0].cpu().numpy() > 0.5).astype(np.uint8)
         else:
             # Always take the first mask, even if YOLO found several
-            mask = masks[0].cpu().numpy() > 0.5
+            mask = (masks[0].cpu().numpy() > 0.5).astype(np.uint8)
             if n_objects > 1:
                 st.info(f"YOLO detected {n_objects} eagles – using the first one automatically.")
